Subject:         Rhino: Problem in NativeJavaMethod    Date:         Tue, 14 Aug 2001 10:23:35 +0200    From:         felix.meschberger@day.com      To:         Norris Boyd <nboyd@atg.com>
Hi Norris,

While working with wrapped Java classes we discovered a problem in
NativeJavaMethod : If the public method to be called is part of a
non-public class, the Sun Java VM throws an IllegalAccessException. This
bug in the Sun VM has been reported as Bug 4071593 to Sun, but has not been
resolved since....

I implemented a circumvention, for which I provide you the patch. I quickly
tested it, and it seems to work.

Regards
Felix