Fixes on arrays, re: bug 121934.  Lots of changed needed to permit arrays to control the types of their children via the information received from the schema.  But the two critical lines actually getting the information on the child types from the schema are commented out, because there are problems with dual hierarchies of schema types that make it not work as well as it should -- also inability to determine whether the schema actually specified a type or just gave a nice default.
This also includes fixes to leaks and other problems discovered reviewing
the array code.  This is the first code capable of inferring and encoding
a multi-dimensional array from the nested arrays of JS, where the array
type dictates.  All my tests still work.

SOAP is not part of the default build.