I found three small but important errors in type identification processes.
Two of them involve the missing classes Struct and AnySimpleType, which I
have been able to fake out.  The third involves the identification of the
type of the array.  When identifying the type of a variant, interface means
that it contains a struct.  but I was using the same method to identify the
type of an array, where interface always means you don't know the type (since
variants show up as interfaces in this case and even nested arrays are
variants).

SOAP is not part of default build.