The problem is that on windows widget can only be positioned to a y coord of 32767. This ends up being about 32 pages. The nsPageFrame's view creates a widget for clipping absolutely positioned content and plugins.
This fix will allow a document to have up to 32 pages of clipping after that it
stops creating the widget. The down side is that any absolutely positioned
content or plugins beyond page 32 will not be clippe din print preview.
Also, fixed a leaked widget
Bug 127263 r=dcone sr=attinasi a=asa