1) Adds two bools to the PrintSettings and then when the PS gets initialized from Prefs or from the Printer they get set. They get unset when the printer name changes.
This is all necessary because the Page Setup Dialog calls
initPrintSettingsFromPrinter each time it is shown. So if you have set values
they will get "written" over the next time the dialog is shown.

2) Removes a lot of platform specific initialization from nsPrintOptionsImpl
and it now uses just the the "generic" XP routines and is now consistent with
the other platforms for initializing the PS.
Bug 167894 r=dcone sr=kin a=roc