fix the junk mail purge code.  when this code was original written, it was modeled after the biff code, using the purge interval as the timer.  (yikes, a 14 day timer!)
now, we fire a timer at a fixed interval, and see if a purge is necessary, and use
the purge interval to decide what messages to delete.

add code to verify that we are only trying to delete messages marked as junk.

we do this because the junk folder might be trash, or there might be non-junk in it
that hasn't been moved, and we don't want to delete that.

r/sr=bienvenu