fix for bug #205877
when you toggle junk status from thread pane icon,
and you've got it set to move on manual mark, assertion and painting problems.

the problem is that ::OnDeleteCompleted() uses the selection to determine which
view indices to call with ::NoteChange().

the junk column icon triggers a delete (or move) and that row might not be selected.

instead, we save of the indices (we can't save keys, as the keys are already removed
before ::OnDeleteCompleted() is called), and use the saved indices when calling
::NoteChange()

thanks to bienvenu for the suggestion.

r/sr=bienvenu