Ugly thunderbird hack. Has no effect on seamonkey builds.
When launching the app via Launch Task (which is the case when starting up for the first time), check to see if the cmd line service
has any arguments (such as mailto urls). If so pass these arguments into OpenWindow instead of using the handler's default arguments.

This fixes the problem where clicking on a mailto url in a browser when thunderbird was NOT already running, caused a blank
compose window where the mailto url was ignored.