Bug 33282       enable external scheme handlers (like aim: and telnet:) in Linux
This allows setting external protocol handlers via a preference setting:
network.protocol-handler.app.<scheme>
This is supposed to be a string value.
Mozilla will first try to interpret this as an absolute path, then as a filename
relative to $MOZILLA_FIVE_HOME, then as a file in $PATH.
If an application is found in one of these places, it will be called and passed
the complete url as first (and only) argument.

r=jesup/bz sr=bz/darin